Update for SeeC 0.27.
This version of SeeC is based on LLVM/Clang 3.8., so the main
differences in the installer (other than the binaries themselves
changing) is that Clang's headers are in a different location (and there
are some additional headers).

This also implements update logic for the installer, so it should
replace the previous 0.26 release of SeeC (and refuse to be installed
over future releases of SeeC).